refactor arrow
