+++ narwhals/_arrow/series.py
@@ -8,13 +8,17 @@ from typing import Literal
 from typing import Sequence
 from typing import overload
 
+from narwhals._arrow.series_cat import ArrowSeriesCatNamespace
+from narwhals._arrow.series_dt import ArrowSeriesDateTimeNamespace
+from narwhals._arrow.series_list import ArrowSeriesListNamespace
+from narwhals._arrow.series_str import ArrowSeriesStringNamespace
 from narwhals._arrow.utils import broadcast_and_extract_native
 from narwhals._arrow.utils import cast_for_truediv
 from narwhals._arrow.utils import floordiv_compat
 from narwhals._arrow.utils import narwhals_to_native_dtype
 from narwhals._arrow.utils import native_to_narwhals_dtype
 from narwhals._arrow.utils import pad_series
-from narwhals._arrow.utils import parse_datetime_format
+from narwhals.typing import CompliantSeries
 from narwhals.utils import Implementation
 from narwhals.utils import generate_temporary_column_name
 from narwhals.utils import import_dtypes_module
@@ -32,7 +36,6 @@ if TYPE_CHECKING:
     from narwhals._arrow.namespace import ArrowNamespace
     from narwhals.dtypes import DType
     from narwhals.utils import Version
-from narwhals.typing import CompliantSeries
 
 
 def maybe_extract_py_scalar(value: Any, return_py_scalar: bool) -> Any:  # noqa: FBT001
@@ -1180,418 +1183,3 @@ class ArrowSeries(CompliantSeries):
     @property
     def list(self: Self) -> ArrowSeriesListNamespace:
         return ArrowSeriesListNamespace(self)
-
-
-class ArrowSeriesDateTimeNamespace:
-    def __init__(self: Self, series: ArrowSeries) -> None:
-        self._compliant_series = series
-
-    def to_string(self: Self, format: str) -> ArrowSeries:  # noqa: A002
-        import pyarrow.compute as pc
-
-        # PyArrow differs from other libraries in that %S also prints out
-        # the fractional part of the second...:'(
-        # https://arrow.apache.org/docs/python/generated/pyarrow.compute.strftime.html
-        format = format.replace("%S.%f", "%S").replace("%S%.f", "%S")
-        return self._compliant_series._from_native_series(
-            pc.strftime(self._compliant_series._native_series, format)
-        )
-
-    def replace_time_zone(self: Self, time_zone: str | None) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        if time_zone is not None:
-            result = pc.assume_timezone(
-                pc.local_timestamp(self._compliant_series._native_series), time_zone
-            )
-        else:
-            result = pc.local_timestamp(self._compliant_series._native_series)
-        return self._compliant_series._from_native_series(result)
-
-    def convert_time_zone(self: Self, time_zone: str) -> ArrowSeries:
-        import pyarrow as pa
-
-        if self._compliant_series.dtype.time_zone is None:  # type: ignore[attr-defined]
-            result = self.replace_time_zone("UTC")._native_series.cast(
-                pa.timestamp(self._compliant_series._native_series.type.unit, time_zone)
-            )
-        else:
-            result = self._compliant_series._native_series.cast(
-                pa.timestamp(self._compliant_series._native_series.type.unit, time_zone)
-            )
-
-        return self._compliant_series._from_native_series(result)
-
-    def timestamp(self: Self, time_unit: Literal["ns", "us", "ms"] = "us") -> ArrowSeries:
-        import pyarrow as pa
-        import pyarrow.compute as pc
-
-        s = self._compliant_series._native_series
-        dtype = self._compliant_series.dtype
-        dtypes = import_dtypes_module(self._compliant_series._version)
-        if dtype == dtypes.Datetime:
-            unit = dtype.time_unit  # type: ignore[attr-defined]
-            s_cast = s.cast(pa.int64())
-            if unit == "ns":
-                if time_unit == "ns":
-                    result = s_cast
-                elif time_unit == "us":
-                    result = floordiv_compat(s_cast, 1_000)
-                else:
-                    result = floordiv_compat(s_cast, 1_000_000)
-            elif unit == "us":
-                if time_unit == "ns":
-                    result = pc.multiply(s_cast, 1_000)
-                elif time_unit == "us":
-                    result = s_cast
-                else:
-                    result = floordiv_compat(s_cast, 1_000)
-            elif unit == "ms":
-                if time_unit == "ns":
-                    result = pc.multiply(s_cast, 1_000_000)
-                elif time_unit == "us":
-                    result = pc.multiply(s_cast, 1_000)
-                else:
-                    result = s_cast
-            elif unit == "s":
-                if time_unit == "ns":
-                    result = pc.multiply(s_cast, 1_000_000_000)
-                elif time_unit == "us":
-                    result = pc.multiply(s_cast, 1_000_000)
-                else:
-                    result = pc.multiply(s_cast, 1_000)
-            else:  # pragma: no cover
-                msg = f"unexpected time unit {unit}, please report an issue at https://github.com/narwhals-dev/narwhals"
-                raise AssertionError(msg)
-        elif dtype == dtypes.Date:
-            time_s = pc.multiply(s.cast(pa.int32()), 86400)
-            if time_unit == "ns":
-                result = pc.multiply(time_s, 1_000_000_000)
-            elif time_unit == "us":
-                result = pc.multiply(time_s, 1_000_000)
-            else:
-                result = pc.multiply(time_s, 1_000)
-        else:
-            msg = "Input should be either of Date or Datetime type"
-            raise TypeError(msg)
-        return self._compliant_series._from_native_series(result)
-
-    def date(self: Self) -> ArrowSeries:
-        import pyarrow as pa
-
-        return self._compliant_series._from_native_series(
-            self._compliant_series._native_series.cast(pa.date32())
-        )
-
-    def year(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.year(self._compliant_series._native_series)
-        )
-
-    def month(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.month(self._compliant_series._native_series)
-        )
-
-    def day(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.day(self._compliant_series._native_series)
-        )
-
-    def hour(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.hour(self._compliant_series._native_series)
-        )
-
-    def minute(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.minute(self._compliant_series._native_series)
-        )
-
-    def second(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.second(self._compliant_series._native_series)
-        )
-
-    def millisecond(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.millisecond(self._compliant_series._native_series)
-        )
-
-    def microsecond(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        arr = self._compliant_series._native_series
-        result = pc.add(pc.multiply(pc.millisecond(arr), 1000), pc.microsecond(arr))
-
-        return self._compliant_series._from_native_series(result)
-
-    def nanosecond(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        arr = self._compliant_series._native_series
-        result = pc.add(
-            pc.multiply(self.microsecond()._native_series, 1000), pc.nanosecond(arr)
-        )
-        return self._compliant_series._from_native_series(result)
-
-    def ordinal_day(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.day_of_year(self._compliant_series._native_series)
-        )
-
-    def weekday(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.day_of_week(self._compliant_series._native_series, count_from_zero=False)
-        )
-
-    def total_minutes(self: Self) -> ArrowSeries:
-        import pyarrow as pa
-        import pyarrow.compute as pc
-
-        arr = self._compliant_series._native_series
-        unit = arr.type.unit
-
-        unit_to_minutes_factor = {
-            "s": 60,  # seconds
-            "ms": 60 * 1e3,  # milli
-            "us": 60 * 1e6,  # micro
-            "ns": 60 * 1e9,  # nano
-        }
-
-        factor = pa.scalar(unit_to_minutes_factor[unit], type=pa.int64())
-        return self._compliant_series._from_native_series(
-            pc.cast(pc.divide(arr, factor), pa.int64())
-        )
-
-    def total_seconds(self: Self) -> ArrowSeries:
-        import pyarrow as pa
-        import pyarrow.compute as pc
-
-        arr = self._compliant_series._native_series
-        unit = arr.type.unit
-
-        unit_to_seconds_factor = {
-            "s": 1,  # seconds
-            "ms": 1e3,  # milli
-            "us": 1e6,  # micro
-            "ns": 1e9,  # nano
-        }
-        factor = pa.scalar(unit_to_seconds_factor[unit], type=pa.int64())
-
-        return self._compliant_series._from_native_series(
-            pc.cast(pc.divide(arr, factor), pa.int64())
-        )
-
-    def total_milliseconds(self: Self) -> ArrowSeries:
-        import pyarrow as pa
-        import pyarrow.compute as pc
-
-        arr = self._compliant_series._native_series
-        unit = arr.type.unit
-
-        unit_to_milli_factor = {
-            "s": 1e3,  # seconds
-            "ms": 1,  # milli
-            "us": 1e3,  # micro
-            "ns": 1e6,  # nano
-        }
-
-        factor = pa.scalar(unit_to_milli_factor[unit], type=pa.int64())
-
-        if unit == "s":
-            return self._compliant_series._from_native_series(
-                pc.cast(pc.multiply(arr, factor), pa.int64())
-            )
-
-        return self._compliant_series._from_native_series(
-            pc.cast(pc.divide(arr, factor), pa.int64())
-        )
-
-    def total_microseconds(self: Self) -> ArrowSeries:
-        import pyarrow as pa
-        import pyarrow.compute as pc
-
-        arr = self._compliant_series._native_series
-        unit = arr.type.unit
-
-        unit_to_micro_factor = {
-            "s": 1e6,  # seconds
-            "ms": 1e3,  # milli
-            "us": 1,  # micro
-            "ns": 1e3,  # nano
-        }
-
-        factor = pa.scalar(unit_to_micro_factor[unit], type=pa.int64())
-
-        if unit in {"s", "ms"}:
-            return self._compliant_series._from_native_series(
-                pc.cast(pc.multiply(arr, factor), pa.int64())
-            )
-        return self._compliant_series._from_native_series(
-            pc.cast(pc.divide(arr, factor), pa.int64())
-        )
-
-    def total_nanoseconds(self: Self) -> ArrowSeries:
-        import pyarrow as pa
-        import pyarrow.compute as pc
-
-        arr = self._compliant_series._native_series
-        unit = arr.type.unit
-
-        unit_to_nano_factor = {
-            "s": 1e9,  # seconds
-            "ms": 1e6,  # milli
-            "us": 1e3,  # micro
-            "ns": 1,  # nano
-        }
-
-        factor = pa.scalar(unit_to_nano_factor[unit], type=pa.int64())
-
-        return self._compliant_series._from_native_series(
-            pc.cast(pc.multiply(arr, factor), pa.int64())
-        )
-
-
-class ArrowSeriesCatNamespace:
-    def __init__(self: Self, series: ArrowSeries) -> None:
-        self._compliant_series = series
-
-    def get_categories(self: Self) -> ArrowSeries:
-        import pyarrow as pa
-
-        ca = self._compliant_series._native_series
-        out = pa.chunked_array(
-            [pa.concat_arrays(x.dictionary for x in ca.chunks).unique()]
-        )
-        return self._compliant_series._from_native_series(out)
-
-
-class ArrowSeriesStringNamespace:
-    def __init__(self: Self, series: ArrowSeries) -> None:
-        self._compliant_series = series
-
-    def len_chars(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.utf8_length(self._compliant_series._native_series)
-        )
-
-    def replace(
-        self: Self, pattern: str, value: str, *, literal: bool, n: int
-    ) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        method = "replace_substring" if literal else "replace_substring_regex"
-        return self._compliant_series._from_native_series(
-            getattr(pc, method)(
-                self._compliant_series._native_series,
-                pattern=pattern,
-                replacement=value,
-                max_replacements=n,
-            )
-        )
-
-    def replace_all(
-        self: Self, pattern: str, value: str, *, literal: bool
-    ) -> ArrowSeries:
-        return self.replace(pattern, value, literal=literal, n=-1)
-
-    def strip_chars(self: Self, characters: str | None) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        whitespace = " \t\n\r\v\f"
-        return self._compliant_series._from_native_series(
-            pc.utf8_trim(
-                self._compliant_series._native_series,
-                characters or whitespace,
-            )
-        )
-
-    def starts_with(self: Self, prefix: str) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.equal(self.slice(0, len(prefix))._native_series, prefix)
-        )
-
-    def ends_with(self: Self, suffix: str) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.equal(self.slice(-len(suffix), None)._native_series, suffix)
-        )
-
-    def contains(self: Self, pattern: str, *, literal: bool) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        check_func = pc.match_substring if literal else pc.match_substring_regex
-        return self._compliant_series._from_native_series(
-            check_func(self._compliant_series._native_series, pattern)
-        )
-
-    def slice(self: Self, offset: int, length: int | None) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        stop = offset + length if length is not None else None
-        return self._compliant_series._from_native_series(
-            pc.utf8_slice_codeunits(
-                self._compliant_series._native_series, start=offset, stop=stop
-            ),
-        )
-
-    def to_datetime(self: Self, format: str | None) -> ArrowSeries:  # noqa: A002
-        import pyarrow.compute as pc
-
-        if format is None:
-            format = parse_datetime_format(self._compliant_series._native_series)
-
-        return self._compliant_series._from_native_series(
-            pc.strptime(self._compliant_series._native_series, format=format, unit="us")
-        )
-
-    def to_uppercase(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.utf8_upper(self._compliant_series._native_series),
-        )
-
-    def to_lowercase(self: Self) -> ArrowSeries:
-        import pyarrow.compute as pc
-
-        return self._compliant_series._from_native_series(
-            pc.utf8_lower(self._compliant_series._native_series),
-        )
-
-
-class ArrowSeriesListNamespace:
-    def __init__(self: Self, series: ArrowSeries) -> None:
-        self._arrow_series = series
-
-    def len(self: Self) -> ArrowSeries:
-        import pyarrow as pa
-        import pyarrow.compute as pc
-
-        return self._arrow_series._from_native_series(
-            pc.cast(pc.list_value_length(self._arrow_series._native_series), pa.uint32())
-        )
